# public/docs/swagger.yaml
...
swagger: '2.0'
...
info:
  version: "0.5.0"
  title: all drivers
  description: |
    DESCRIPTION                                           
    for urls with authentication, you need to send authKey and phone variables too               
    for example (url POST /orders) :                                      
    {                             
      phone:'89888888888',                        
      authKey:'ad03dj0as9dja',                                  
      orders: [order1, order2]                              
    }
    models in DB
host: '46.101.157.129'
...
paths:
...
  /orders:
    get:
      summary: OK
      description: Gets ?active `Order` objects for authenticated user
      responses:
        200:
          description: OK
          schema:
            properties:
              msg:
                type: array
                items:
                  $ref: '#/definitions/Order'
    post:
      summary: OK
      description: |
        Adds list of `Order` object and returns new order ids.             
        THIS CASE NEEDS SOME MORE DETAILED VIEW:                          
            * when you need it in accurate time
            * when you need it as soon as possible
            # * when you need it now, but not in less than 5 minutes (you can wait a little)
        ATTENTION! you need to pass an ARRAY of orders in this method!!
        the proper object will be                                 
        { orders: [order1, order2] }
      parameters:
        - name: phone
          in: formData
          required: true
          description: user phone
          type: string
        - name: orders
          in: formData
          required: true
          description: array of orders
          schema:
            type: array
            items:
              type: object
              properties:
                location:
                  description: location of client
                  required: true
                  $ref: '#/definitions/Address'
                destination:
                  description: client's destination
                  $ref: '#/definitions/Address'
                  required: false
                asSoonAsPossible:
                  type: boolean
                orderDate:
                  required: true
                  type: string
                  format: date
                  description: time, when client needs a car. if asSoonAsPossible is false/0 you need to specify this parameter
                clientPhone:
                  type: string
                driverDetails:
                  $ref: '#/definitions/DriverDetails'

      responses:
        200:
          description: OK
        500:
          description: Error
          schema:
            type: object
            items:
              $ref: '#/definitions/Error'
...
  /orders/all:
    get:
      summary: OK
      description: Gets all `Order`s by period of time (if specified)
      parameters:
        - name: date1
          in: query
          description: format??
          type: string
          format: date
        - name: date2
          in: query
          description: format??
          type: string
          format: date
      responses:
        200:
          description: OK
...
  /orders/drivers:
    get:
      summary: not implemented
      description: gets list of drivers, who are ready to pick order
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              msg:
                type: object
                properties:
                  drivers:
                    type: array
                    items:
                      $ref: '#/definitions/Driver'
    patch:
      summary: not implemented
      description: Attaches chosen `Driver` to `Order`. Needs user authentication
      parameters:
        - name: orderId
          type: string
          in: body
          description: orderId
        - name: drivers
          in: body
          description: array of driver Ids
          schema:
            type: array
            items:
              name: driverId
              type: string
      responses:
        200:
          description: OK
...
  /drivers:
    post:
      summary: OK
      description: Adds `Driver` object
      parameters:
        - name: name
          in: formData
          description: driverName
          type: string
        - name: categories
          in: body
          description: |
            { A: 1, B: 1 , C: 0, D: 0, E: 1 }
          schema:
            properties:
              A:
                type: number
        - name: internationalLicence
          in: formData
          type: number
          description: 1 - has international driving licence, 0 - not
        - name: gender
          in: formData
          type: number
          description: 1 - male, 0 - female
        - name: isOfficial
          in: formData
          type: number
          description: 1 - wears official clothes. 0 - not
        - name: appearance
          in: formData
          type: string
          description: comment to Appearance
          required: false
        - name: isSmoking
          in: formData
          type: number
          description: 1 - has smoking habit. 0 - does not
      responses:
        200:
          description: OK
        500:
          description: Error
          schema:
            type: object
            items:
              $ref: '#/definitions/Error'
...
definitions:
  Error:
    type: object
    properties:
      code:
        type: string
      message:
        type: string
...
  Order:
    type: object
    properties:
      userId:
        type: string
      drivers:
        type: array
        items:
          type: string
      status:
        type: number
      added:
        type: string
        format: date
      location:
        type: object
      destination:
        type: object
      asSoonAsPossible:
        type: boolean
      orderDate:
        type: string
        format: date
        description: if asSoonAsPossible is false/0 you need to specify this parameter
      clientPhone:
        type: string
      driverDetails:
        $ref: '#/definitions/DriverDetails'
      options:
        type: object
        description: // smoking and all such stuff. need more info from manager to fill this object
  Driver:
    type: object
    properties:
      name:
        type: string
      categories:
        type: object
      internationalLicence:
        type: boolean
      gender:
        type: boolean
      isOfficial:
        type: boolean
      appearance:
        type: string
        description: // comment to Appearance                  
      isSmoking:
        type: boolean
  DriverDetails:
    type: object
    properties:
      name:
        type: string
      categories:
        type: object
      internationalLicence:
        type: boolean
      gender:
        type: boolean
      isOfficial:
        type: boolean
      appearance:
        type: string
        description: // comment to Appearance                  
      isSmoking:
        type: boolean
...
  Address:
    type: object
    properties:
      center:
        type: array
        items:
          type: number
          format: float
      name:
        type: string